+++ core/src/lib.rs
@@ -27,6 +27,7 @@ mod character;
 mod collect;
 pub mod color_transform;
 pub mod context;
+pub mod context_menu;
 mod drawing;
 mod ecma_conversions;
 pub mod events;
@@ -50,9 +51,9 @@ pub mod backend;
 pub mod config;
 pub mod external;
 
-pub use avm1::ContextMenuInfo;
-pub use avm1::ContextMenuItemInfo;
 pub use chrono;
+pub use context_menu::ContextMenuInfo;
+pub use context_menu::ContextMenuItemInfo;
 pub use events::PlayerEvent;
 pub use indexmap;
 pub use player::Player;
